requirements.txt
inquirer==2.7.0
PyYAML==5.4.1
requests==2.25.1
